+++ run/lib/logger.js
@@ -7,20 +7,19 @@ const logger = createLogger({
     transports: []
 });
 
-if (process.env.DATADOG_API_KEY) {
-    const httpTransportOptions = {
-        host: 'http-intake.logs.datadoghq.eu',
-        path: `/api/v2/logs?dd-api-key=${process.env.DATADOG_API_KEY}&ddsource=nodejs&service=ethernal`,
-        ssl: true
-    };
-    const transport = new transports.Http(httpTransportOptions);
-    logger.add(transport);
-    logger.exceptions.handle(transport);
+let newrelicWinstonFormatter, transport;
+
+if (process.env.NODE_ENV == 'production') {
+    const newrelicFormatter = require('@newrelic/winston-enricher');
+    newrelicWinstonFormatter = newrelicFormatter(require('winston'));
 }
-else {
-    const transport = new transports.Console({ format: format.combine(format.colorize(), format.simple()) });
-    logger.add(transport);
-    logger.exceptions.handle(transport);
-}
+
+if (newrelicWinstonFormatter)
+    transport = new transports.Console({ format: format.combine(format.colorize(), format.simple(), newrelicWinstonFormatter()) });
+else
+    transport = new transports.Console({ format: format.combine(format.colorize(), format.simple()) });
+
+logger.add(transport);
+logger.exceptions.handle(transport);
 
 module.exports = logger;