+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.4.2] - 2022-12-30
+### Changed
+- Improved CI/CD process to include everything
+
+### Added
+- Newrelic monitoring
+
 ## [3.4.1] - 2022-12-29
 ### Fixed
 - Frontend auth